+++ integration_test/app_test.dart
@@ -6,10 +6,10 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
 import 'package:logging/logging.dart';
 import 'package:todo/di/injector.dart';
-import 'package:todo/features/tasks_overview/view/main_screen.dart';
-import 'package:todo/features/tasks_overview/widgets/app_bar.dart';
 import 'package:todo/main.dart' as app;
-import 'package:todo/utils/key_widget.dart';
+import 'package:todo/src/core/utils/key_widget.dart';
+import 'package:todo/src/feature/tasks_overview/view/main_screen.dart';
+import 'package:todo/src/feature/tasks_overview/widgets/app_bar.dart';
 
 final testLogger = Logger('[Test]');
