+++ lib/data/models/tasks.dart
@@ -1,6 +1,6 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:todo/data/models/only_task.dart';
-import 'package:todo/utils/serializable.dart';
+import 'package:todo/src/core/utils/serializable.dart';
 
 part 'tasks.freezed.dart';
 part 'tasks.g.dart';
