+++ lib/src/core/utils/serializable.dart
