+++ lib/src/core/utils/types.dart
