+++ lib/src/feature/edit_add_task/widgets/delete_icon_button.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
-import 'package:todo/features/edit_add_task/bloc/bloc.dart';
-import 'package:todo/features/edit_add_task/view/edit_add_task_screen.dart';
-import 'package:todo/styling/text_styles.dart';
-import 'package:todo/utils/extensions.dart';
+import 'package:todo/src/core/constant/styling/text_styles.dart';
+import 'package:todo/src/core/utils/extensions.dart';
+import 'package:todo/src/feature/edit_add_task/bloc/bloc.dart';
+import 'package:todo/src/feature/edit_add_task/view/edit_add_task_screen.dart';
 
 class DeleteIconButtonWidget extends StatelessWidget {
   const DeleteIconButtonWidget({
