+++ lib/src/feature/edit_add_task/widgets/task_title_card_text_field.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
-import 'package:todo/features/edit_add_task/bloc/bloc.dart';
-import 'package:todo/features/edit_add_task/view/edit_add_task_screen.dart';
+import 'package:todo/src/feature/edit_add_task/bloc/bloc.dart';
+import 'package:todo/src/feature/edit_add_task/view/edit_add_task_screen.dart';
 
 class TextFieldTaskNameWidget extends StatelessWidget {
   const TextFieldTaskNameWidget({
