+++ lib/src/feature/tasks_overview/view/main_screen.dart
@@ -3,21 +3,21 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:todo/data/models/only_task.dart';
-import 'package:todo/features/tasks_overview/bloc/bloc.dart';
-import 'package:todo/features/tasks_overview/filter_tasks.dart';
-import 'package:todo/features/tasks_overview/widgets/app_bar.dart';
-import 'package:todo/features/tasks_overview/widgets/check_box.dart';
-import 'package:todo/features/tasks_overview/widgets/dismiss_container.dart';
-import 'package:todo/features/tasks_overview/widgets/edit_icon.dart';
-import 'package:todo/features/tasks_overview/widgets/priority_icon.dart';
-import 'package:todo/features/tasks_overview/widgets/task_title.dart';
-import 'package:todo/features/tasks_overview/widgets/tasks_card.dart';
 import 'package:todo/navigation/cubit_navigation/navigation_cubit.dart';
+import 'package:todo/src/core/constant/styling/colors.dart';
+import 'package:todo/src/core/utils/extensions.dart';
+import 'package:todo/src/core/utils/key_widget.dart';
+import 'package:todo/src/core/utils/priority.dart';
 import 'package:todo/src/core/utils/refined_logger.dart';
-import 'package:todo/styling/colors.dart';
-import 'package:todo/utils/extensions.dart';
-import 'package:todo/utils/key_widget.dart';
-import 'package:todo/utils/priority.dart';
+import 'package:todo/src/feature/tasks_overview/bloc/bloc.dart';
+import 'package:todo/src/feature/tasks_overview/filter_tasks.dart';
+import 'package:todo/src/feature/tasks_overview/widgets/app_bar.dart';
+import 'package:todo/src/feature/tasks_overview/widgets/check_box.dart';
+import 'package:todo/src/feature/tasks_overview/widgets/dismiss_container.dart';
+import 'package:todo/src/feature/tasks_overview/widgets/edit_icon.dart';
+import 'package:todo/src/feature/tasks_overview/widgets/priority_icon.dart';
+import 'package:todo/src/feature/tasks_overview/widgets/task_title.dart';
+import 'package:todo/src/feature/tasks_overview/widgets/tasks_card.dart';
 
 class MainScreenConfigure {
   static const paddingH24 = EdgeInsets.symmetric(horizontal: 24);
