--- lib/styling/assets_name.dart
@@ -1,4 +1,0 @@
-class AssetsNameFile {
-  static const lowPrioritySvg = 'assets/icons/low_priority.svg';
-  static const highPrioritySvg = 'assets/icons/high_priority.svg';
-}
