+++ test/add_edit_task_bloc_test.dart
@@ -11,9 +11,9 @@ import 'package:todo/domain/revision_repository.dart';
 import 'package:todo/domain/task.dart';
 import 'package:todo/domain/task_repository.dart';
 import 'package:todo/domain/tasks_list.dart';
-import 'package:todo/features/edit_add_task/bloc/bloc.dart';
-import 'package:todo/utils/extensions.dart';
-import 'package:todo/utils/priority.dart';
+import 'package:todo/src/core/utils/extensions.dart';
+import 'package:todo/src/core/utils/priority.dart';
+import 'package:todo/src/feature/edit_add_task/bloc/bloc.dart';
 
 class MockTasksRepository extends Mock implements TasksRepository {}
